pyproject.toml
[project]
name = "acdemic-analyser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "asyncio>=3.4.3",
    "autogen-agentchat>=0.5.7",
    "autogen-core>=0.5.7",
    "autogen-ext>=0.5.7",
    "autogenstudio>=0.4.2.2",
    "dotenv>=0.9.9",
    "openai>=1.97.1",
    "pillow>=11.3.0",
    "streamlit>=1.47.1",
    "tiktoken>=0.9.0",
]
